+++ features/column/components/config/date-config-editor.tsx
@@ -7,7 +7,7 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Checkbox } from '@/components/ui/checkbox';
 import { Button } from '@/components/ui/button';
-import { DatePickerCalendar } from '@/components/ui/date-picker-calendar';
+import { DatePickerCalendar } from '@/components/shared/date-picker-calendar';
 import {
   Popover,
   PopoverContent,
