+++ features/table/components/cells/date-cell.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import { format } from 'date-fns';
 import { ja } from 'date-fns/locale';
 import { Button } from '@/components/ui/button';
-import { DatePickerCalendar } from '@/components/ui/date-picker-calendar';
+import { DatePickerCalendar } from '@/components/shared/date-picker-calendar';
 import {
   Popover,
   PopoverContent,
